Added the laser rifle

+++ Assets/Scripts/CombatManager.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class CombatManager : MonoBehaviour
 {
-    private int health;
+    private float health;
     // Start is called before the first frame update
     void Start()
     {
@@ -17,7 +17,7 @@ public class CombatManager : MonoBehaviour
         Debug.Log(health + "Health of enemy");
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (health - damage <= 0)
         {
